Add a sanity check for essential functions
EmbeddedServerController relies on the following functions:
* `proc_open` to start the server
* `fsockopen` to check if the server is running
* `shell_exec` to kill the server

The constructor now checks for availability of these functions and throws
a `RuntimeException` if any of these functions is not available.

This behavior is tested in `MissingFunctionsTest.php` which needs to be executed
separate from the normal unittests, because these functions must be disabled for the test to work. See `.gitlab-ci.yml` for informations on how to execute this test.